+++ src/main/java/org/folio/circulation/StoreLoanAndItem.java
@@ -1,7 +1,7 @@
 package org.folio.circulation;
 
 import static java.util.concurrent.CompletableFuture.completedFuture;
-import static org.folio.circulation.infrastructure.storage.ItemRepository.noLocationMaterialTypeAndLoanTypeInstance;
+import static org.folio.circulation.infrastructure.storage.inventory.ItemRepository.noLocationMaterialTypeAndLoanTypeInstance;
 import static org.folio.circulation.support.Result.succeeded;
 import static org.folio.circulation.support.ResultBinding.mapResult;
 
@@ -10,7 +10,7 @@ import org.folio.circulation.domain.Loan;
 import org.folio.circulation.infrastructure.storage.LoanRepository;
 import org.folio.circulation.resources.context.RenewalContext;
 import org.folio.circulation.support.Clients;
-import org.folio.circulation.infrastructure.storage.ItemRepository;
+import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
 import org.folio.circulation.support.Result;
 import org.folio.circulation.support.http.client.Response;
