+++ src/main/java/org/folio/circulation/infrastructure/storage/PatronActionSessionRepository.java
@@ -30,6 +30,7 @@ import org.folio.circulation.domain.notice.session.PatronSessionRecord;
 import org.folio.circulation.domain.Location;
 import org.folio.circulation.domain.MultipleRecords;
 import org.folio.circulation.domain.User;
+import org.folio.circulation.infrastructure.storage.inventory.LocationRepository;
 import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.CollectionResourceClient;
 import org.folio.circulation.support.Result;
