+++ src/main/java/org/folio/circulation/infrastructure/storage/inventory/MaterialTypeRepository.java
@@ -1,4 +1,4 @@
-package org.folio.circulation.infrastructure.storage;
+package org.folio.circulation.infrastructure.storage.inventory;
 
 import static java.util.Objects.isNull;
 import static java.util.function.Function.identity;
