+++ src/main/java/org/folio/circulation/resources/CheckOutByBarcodeResource.java
@@ -48,7 +48,7 @@ import org.folio.circulation.domain.validation.ServicePointOfCheckoutPresentVali
 import org.folio.circulation.services.EventPublisher;
 import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.ClockManager;
-import org.folio.circulation.infrastructure.storage.ItemRepository;
+import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
 import org.folio.circulation.support.Result;
 import org.folio.circulation.support.RouteRegistration;
 import org.folio.circulation.support.ValidationErrorFailure;
