+++ src/main/java/org/folio/circulation/resources/ItemsInTransitResource.java
@@ -26,7 +26,7 @@ import org.folio.circulation.domain.ItemsReportFetcher;
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.domain.MultipleRecords;
 import org.folio.circulation.infrastructure.storage.PatronGroupRepository;
-import org.folio.circulation.infrastructure.storage.ReportRepository;
+import org.folio.circulation.infrastructure.storage.inventory.ItemReportRepository;
 import org.folio.circulation.domain.Request;
 import org.folio.circulation.domain.ServicePoint;
 import org.folio.circulation.infrastructure.storage.ServicePointRepository;
@@ -35,7 +35,7 @@ import org.folio.circulation.domain.representations.ItemReportRepresentation;
 import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.FindWithMultipleCqlIndexValues;
 import org.folio.circulation.support.GetManyRecordsClient;
-import org.folio.circulation.infrastructure.storage.ItemRepository;
+import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
 import org.folio.circulation.support.Result;
 import org.folio.circulation.support.RouteRegistration;
 import org.folio.circulation.support.http.client.CqlQuery;
@@ -75,12 +75,12 @@ public class ItemsInTransitResource extends Resource {
     final GetManyRecordsClient requestsStorageClient = clients.requestsStorage();
     final ItemRepository itemRepository = new ItemRepository(clients, true, true, true);
     final ServicePointRepository servicePointRepository = new ServicePointRepository(clients);
-    final ReportRepository reportRepository = new ReportRepository(clients);
+    final ItemReportRepository itemReportRepository = new ItemReportRepository(clients);
     final UserRepository userRepository = new UserRepository(clients);
     final PatronGroupRepository patronGroupRepository = new PatronGroupRepository(clients);
     final Comparator<InTransitReportEntry> sortByCheckinServicePointComparator = sortByCheckinServicePointComparator();
 
-    reportRepository.getAllItemsByField("status.name", IN_TRANSIT.getValue())
+    itemReportRepository.getAllItemsByField("status.name", IN_TRANSIT.getValue())
       .thenComposeAsync(r -> r.after(itemsReportFetcher ->
         fetchItemsRelatedRecords(itemsReportFetcher, itemRepository, servicePointRepository)))
       .thenComposeAsync(r -> r.after(inTransitReportEntries ->
