+++ src/main/java/org/folio/circulation/resources/RequestCollectionResource.java
@@ -36,7 +36,7 @@ import org.folio.circulation.domain.validation.UserManualBlocksValidator;
 import org.folio.circulation.services.EventPublisher;
 import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.FindWithCqlQuery;
-import org.folio.circulation.infrastructure.storage.ItemRepository;
+import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
 import org.folio.circulation.support.http.server.JsonHttpResponse;
 import org.folio.circulation.support.http.server.NoContentResponse;
 import org.folio.circulation.support.http.server.WebContext;
