+++ src/main/java/org/folio/circulation/resources/RequestFromRepresentationService.java
@@ -19,7 +19,7 @@ import org.folio.circulation.infrastructure.storage.UserRepository;
 import org.folio.circulation.domain.validation.ProxyRelationshipValidator;
 import org.folio.circulation.domain.validation.ServicePointPickupLocationValidator;
 import org.folio.circulation.support.BadRequestFailure;
-import org.folio.circulation.infrastructure.storage.ItemRepository;
+import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
 import org.folio.circulation.support.Result;
 
 import io.vertx.core.json.JsonObject;
