+++ src/main/java/org/folio/circulation/resources/renewal/RenewByBarcodeResource.java
@@ -9,7 +9,7 @@ import org.folio.circulation.infrastructure.storage.LoanRepository;
 import org.folio.circulation.infrastructure.storage.UserRepository;
 import org.folio.circulation.storage.SingleOpenLoanByUserAndItemBarcodeFinder;
 import org.folio.circulation.support.Result;
-import org.folio.circulation.infrastructure.storage.ItemRepository;
+import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
 
 import io.vertx.core.http.HttpClient;
 import io.vertx.core.json.JsonObject;
