+++ src/main/java/org/folio/circulation/resources/renewal/RenewByIdResource.java
@@ -13,7 +13,7 @@ import org.folio.circulation.infrastructure.storage.UserRepository;
 import org.folio.circulation.domain.validation.UserNotFoundValidator;
 import org.folio.circulation.storage.ItemByIdInStorageFinder;
 import org.folio.circulation.storage.SingleOpenLoanForItemInStorageFinder;
-import org.folio.circulation.infrastructure.storage.ItemRepository;
+import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
 import org.folio.circulation.support.Result;
 
 import io.vertx.core.http.HttpClient;
