+++ src/main/java/org/folio/circulation/storage/ItemByIdInStorageFinder.java
@@ -7,7 +7,7 @@ import java.util.function.Supplier;
 
 import org.folio.circulation.domain.Item;
 import org.folio.circulation.support.HttpFailure;
-import org.folio.circulation.infrastructure.storage.ItemRepository;
+import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
 import org.folio.circulation.support.Result;
 
 public class ItemByIdInStorageFinder {
